Feat: 장바구니 페이지 - 재설계
navigo 사용을 위한 장바구니 페이지 js로직 재설계

productDetail.js에 장바구니 Js로직 추가
