+++ src/html/cart.html
@@ -75,8 +75,8 @@
     <!-- 공통 헤더 끝 -->
     <!-- main 시작 -->
     <main class="main">
-      <!-- <section class="cart"> -->
-      <!-- <div class="cart__header"><h2>장바구니</h2></div>
+      <!-- <section class="cart">
+        <div class="cart__header"><h2>장바구니</h2></div>
         <div class="cart__container">
           <ul class="cart__list"></ul> -->
 
@@ -107,8 +107,8 @@
                 결제하기
               </button></aa
             >
-          </aside> -->
-      <!-- </div>
+          </aside>
+        </div>
       </section> -->
     </main>
     <!-- main 끝 -->
