+++ src/html/productDetail.html
@@ -7,6 +7,7 @@
     <title>제품 상세페이지</title>
     <!-- scss -->
     <link rel="stylesheet" href="../scss/productDetail/productsDetail.scss" />
+    <link rel="stylesheet" href="../scss/cart/cart.scss" />
     <!-- js -->
     <script
       type="module"
